requirements.txt
gradio>=5.0.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
langgraph>=0.2.56
torch>=2.3.0
sentence-transformers>=3.0.1
numpy>=1.26.4